Remove unnecessary rails files
