--- config/secrets.yml
@@ -1,22 +1,0 @@
-# Be sure to restart your server when you modify this file.
-
-# Your secret key is used for verifying the integrity of signed cookies.
-# If you change this key, all old signed cookies will become invalid!
-
-# Make sure the secret is at least 30 characters and all random,
-# no regular words or you'll be exposed to dictionary attacks.
-# You can use `rake secret` to generate a secure secret key.
-
-# Make sure the secrets in this file are kept private
-# if you're sharing your code publicly.
-
-development:
-  secret_key_base: 5dd0b549087fccdc3d6200d819f35fc4dc917ce8b7b53ce158443ea953b9b2af96d21561d3e5a5e16c6f176d8332bddeaada4b54329f3dbb2f995d628928c428
-
-test:
-  secret_key_base: 717641b73c42c9e1313fd3323e620bc14f18caa6c2762f417ded2017aaa3db1120ee88b6244fd33ec0321976b6f0dc3b10330c494f4576fa3e35815b99535cc3
-
-# Do not keep production secrets in the repository,
-# instead read values from the environment.
-production:
-  secret_key_base: <%= ENV["SECRET_KEY_BASE"] %>
